[refactor:frontend/folder-structure] - refactoring the project folder structure
