+++ client/src/app/product/product.type.ts
