+++ client/src/app/stock-history/index.tsx
