+++ client/src/app/stock-history/stock.type.ts
